Changed reference of siren recording to a gun shot
The previous file 40722-8-0-4.wav was a siren (second number in the name is an 8 which corresponds to a siren). This is an error because a gun shot file was expected--so I changed it to 46656-6-0-0.wav

"The name takes the following format: [fsID]-[classID]-[occurrenceID]-[sliceID].wav"
See https://serv.cusp.nyu.edu/projects/urbansounddataset/urbansound8k.html for all details regarding the file name format.